Split off SliceIndex to it's own module
